src/services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch
torchvision
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
